--- samples/corda/corda-simple-application/clients/src/main/resources/config/credentials/remote-network-config.json.template
@@ -1,26 +1,0 @@
-{
-  "network1": {
-    "type": "fabric",
-    "relayEndpoint": "localhost:9080",
-    "channelName": "mychannel",
-    "chaincode": "simpleassettransfer"
-  },
-  "network2": {
-    "type": "fabric",
-    "relayEndpoint": "localhost:9083",
-    "channelName": "mychannel",
-    "chaincode": "simpleassettransfer"
-  },
-  "Corda_Network": {
-    "type": "corda",
-    "relayEndpoint": "localhost:9081",
-    "partyEndPoint": "localhost:30006",
-    "flowPackage": "com.cordaSimpleApplication.flow"
-  },
-  "Corda_Network2": {
-    "type": "corda",
-    "relayEndpoint": "localhost:9082",
-    "partyEndPoint": "localhost:10006",
-    "flowPackage": "com.cordaSimpleApplication.flow"
-  }
-}
